requirements.txt
streamlit==1.50.0
python-dotenv==1.0.0
openai==1.108.1
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-pinecone==0.2.12
langchain-community==0.3.30
pinecone-client==3.2.2
pydantic==2.10.6
tiktoken==0.9.0